updated Second time

+++ Backend/controller/Section.js
@@ -1,6 +1,6 @@
 const Section = require("../models/Section")
 const Course = require("../models/Course") 
-
+const SubSection = require("../models/SubSection")
 exports.createSection = async (req , res)=>{
     try{
         // data fetch
@@ -80,13 +80,44 @@ exports.deleteSection = async(req , res)=>{
     try{
 
         // get id
-      const {sectionId} = req.body ;
+      const {sectionId , courseId} = req.body ;
+      await Course.findByIdAndUpdate(courseId , {
+          $pull:{
+              courseContent : sectionId ,
+          }
+      })
         // use find by id and delete
-        await Section.findByIdAndDelete(sectionId)
-        // 
+       const section =  await Section.findByIdAndDelete(sectionId)
+
+       console.log(courseId , sectionId)
+
+       if(!section){
+           return res.status(404).json({
+               success : false ,
+               message : "section not found"
+           })
+       }
+        // delete subsection
+
+        await SubSection.deleteMany({_id:{$in:section.subSection}}) ;
+
+        await Section.findByIdAndDelete(sectionId) ;
+
+        const course =await (await Course.findById(courseId)).populated({
+            path:"courseContent" ,
+            populate :{
+                path :" subSection"
+            }
+        }).exec() ;
+        
+        
+
+
+
+
         return res.status(200).json({
             success : true,
-            message : "Deleted succesfully"
+            message : "section Deleted succesfully"
           })
 
     }catch(err){
